refactor(refs T33194, T19759): Restore state to have less changes
